--- a/tox.ini
+++ b/tox.ini
@@ -18,7 +18,6 @@ commands =
       --cov-report term \
       --cov-report xml \
       {posargs}
-    coverage report
 
 [testenv:lint]
 skip_install = true