Command-line interfaces value speed, and unused imports can slow things down a lot.

We can use `python -X importtime` and tuna to identify bottlenecks: https://medium.com/alan/how-we-improved-our-python-backend-start-up-time-2c33cd4873c8

For example:

```sh
python3.13 -X importtime -c 'import prettytable' 2> import.log
tuna import.log
```

We're using Python 3.13.0a3 here, because it already includes some optimisations of its own.

# Baseline

Starting on `main`, `prettytable`: 0.027 s (87.8%)

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/8c699317-63f7-4039-afbe-e2f861b81328">

And let's begin by zooming into `prettytable.prettytable`: 0.010s (32.8%)

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/5d6c82c6-4618-4c57-98a4-9413f303f339">

# Lazy import json

`json` is a good first target. Taking 0.001 s (2.7%), it's only used by one function to load from JSON, and another to save. If doing plain text tables (or HTML etc), you won't use them.
Let's move the import into those two functions where it's actually needed:

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/2c40515f-5875-4ec0-bc0f-3b223a943fac">

Result: `prettytable.prettytable`: 0.009 s (30.3%)

# Lazy import random

`random` is only used by a function to set random styles, and is mostly for fun. Lazy import it to save 0.001 s (1.8%):

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/e58b93e7-c5b6-49e4-a03a-b9069303701e">

Result: `prettytable.prettytable` 0.009 s (28.5%)

# Lazy import csv

Like `json`, `csv` is only used for loading or saving CSV and takes 0.003 s (10.8%).

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/8d7402ad-60d8-4667-bcc7-e705f1c3b44d">

Result: `prettytable.prettytable` 0.008 s (27.1%)

# Lazy import copy

Used in three places: a function to return a copy of a table. Is this used much? And also in other functions, themselves used when getting a number of string, HTML, CSV, JSON, Latex versions of a table. That's quite a few, so will likely end up being imported anyway, but it's a simple change and we may save 0.001 s (2.7%) in a few places.

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/c8a98ee6-178b-4ecd-9ca2-88a79684fa2b">

Result: `prettytable.prettytable` 0.008 s (25.9%)

# Lazy import textwrap

Only used in one function when getting a string version of a table. I think a common operation, but again, a quick saving for those cases when not.

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/5a8dba4f-165c-4636-aeed-84ea46a19b9c">

Result: `prettytable.prettytable` 0.007 s (24.6%)

# Lazy import wcwidth

Called by one function, but itself called several times, mostly for getting string tables? Well, saving of 0.001 s (2.2%) if not needed.


Result: `prettytable.prettytable` 0.007 s (23.0%)

# Lazy import html.escape

Only used when getting HTML versions of tables: 0.001 s (2.3%)

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/baf83b19-447a-4085-9e76-6615b5318292">

Result: `prettytable.prettytable` 0.007 s (22.5%)


# Replace math.floor with int

For positive inputs `math.floor` and `int` give the same results, we can avoid importing `math` to save a (rounded) 0.000 s (0.9%). We were calling `int` on the result of `math.floor` anyway!

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/a26351a6-d32a-48da-a3b3-6e24739eafb2">

Result: `prettytable.prettytable` 0.006 s (21.6%)

# Lazy import importlib.metadata

Finally, let's step out back to `prettytable`:

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/31ffe7c1-f3da-4e3e-adf2-2c77e12bee5b">

`importlib.metadata` takes a big chunk of the original: 0.015 s (50.9%)!

We only use this to set `__version__` on the off-chance someone might access it:

```python
import importlib.metadata

__version__ = importlib.metadata.version(__name__)
```

We can instead only import when `__version__` is accessed:

```python
def __getattr__(name: str) -> Any:
    if name == "__version__":
        import importlib.metadata

        return importlib.metadata.version(__name__)

    msg = f"module '{__name__}' has no attribute '{name}'"
    raise AttributeError(msg)
``` 

<img width="1582" alt="image" src="https://github.com/jazzband/prettytable/assets/1324225/a9337b42-c658-4bd2-8ff1-1e686e7f103e">

Result: `prettytable` 0.007 s (65.1%).

---

Compare with our original `prettytable`: 0.027 s (87.8%), that's a big saving! 

This is an idealised result and in reality a few of these will be imported, based on the use case. Still, those that do apply will help CLIs.

A quick look at the remaining "big" ones:

* `typing` 0.003 s (24.3%) - not easy to remove without removing type hints entirely. The stdlib has already optimised `typing` for 3.13, so already better than before.
* `re` 0.002 s (18.5%) - used to compile a regex to save time later. Used before calculating widths. Could potentially be replaced with some `str.replace` calls. Something to consider another time.
* `html.parser.HTMLParser` 0.002 s (17.0%) - used to subclass. Could potentially be moved and only imported/created when needing to parse those things, but could technically change the public API. Something to consider another time.

But we've achieved enough here for now!